NL classifier starts disabled by default
when the ecm_classifier_nl module starts, it will start in a
disabled state.

this is needed because we don't want to prevent acceleration
when StreamBoost is not loaded.  when StreamBoost starts, it
will be responsible for enabling the classifier and will also
be reponsible for disabling it when it stops.

Change-Id: Id5f9c34c3f23288bd34dd97e6172d119254e45a9
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>