[ecm_classifier_nl] issue NL close only on refs=0
the NL classifier should only issue a msg_CLOSE when it receives
a conntrack close event, but since these events are propagated to
the classifier as a deref, we check instead for refs becoming zero
as the trigger for the message.

previously, a close would be issued on reception of an
NSS_SYNC_REASON_DESTROY, but the meaning of destroy from an NSS
sync is not the same as a destroy from conntrack.  rather, a
destroy from NSS means that the offload rule associated with the
connection was destroyed, but the connection itself may still be
open and passing traffic.

Change-Id: Ia71d15cc5904fda1b6337ad633d46250909d49c9
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>