[qca-nss-ecm] Bridge code should use bridge port for 'in'.
Using skb->skb_iif to obtain 'in' device is wrong since this will not work
where there is a heirarchy of interfaces such as a eth0 and eth0.10 vlan as a bridge port.
Instead use port lookup API to locate the actual port device that reaches the source MAC
which will be equivalent for the true 'in' device.

Change-Id: I84901bbf155fd475807ec85497a0bbd902a4b910
Signed-off-by: Gareth Williams <garethw@codeaurora.org>