[qca-nss-ecm] Issue close upon a connection removed event rather than on classifier final deref
Final deref is not a good place since the connection upon which the classifier depends so as to be
able to construct the tuple to send in a close event to the netlink is no longer around.
Instead responding to a db removed event the classifier can use the connection state to get the tuple
and send the close event.

Change-Id: I7b8f336fa20dc52ff38d10c5a3614cf325e218a8
Signed-off-by: Gareth Williams <garethw@codeaurora.org>