Revert "nl classifer: do not action QOS if skb->priority is nonzero"
This reverts commit d4ee0090d591db12b52d74d690569173ef8a5cf9.

Since the nl classifier does not action QOS in the case that skb
priority is nonzero, the front end allows the next lesser priority
classifier, in this case dscp, a chance to mark.  However, in lab
tests, the dscp classifier does not set the desired mark for both
directions.  In the test, we see dscp correctly mark the flow_qos_tag,
but not the return_qos_tag, so it remains zero.  This breaks
classification.

Due to this unresolved interaction issue with other classifiers, we
need to revert this change so that streamboost always actions a QOS
tag if it has one.