NL classifier: query ct mark on db entry create
When an DB connection entry is created, query conntrack for the
current mark.

This closes a window of time that can exist between conntrack
connection mark updates and notifications and ecm_db conntrack
instance creation.

If the nl_classifier instance can miss conntrack mark update events
for a connection that does not yet exist in ecm_db, we must ensure
that such marks are read on DB connection entry creation.

Change-Id: I09c930192ee077f53b3ea7184b4600e1ae6c3e7c
Signed-off-by: Ben Menchaca <bmenchac@codeaurora.org>