reduce nl msg buffer size, check msg return code
reduced the message buffer size of netlink messages to prevent
memory waste, which becomes a problem when many messages are
issued simultaneously.

in empirical testing, we saw that we could handle around 7
outstanding messages before dropping and with this change we
tested up to 80 with no failures.

this change also adds return code checking and adds a return
code to the nl msg send function.

CRs-Fixed: 695456
Change-Id: If3c36ea54ff56ce2035ccf3f0f70b1bba27f8d72
Signed-off-by: Nicolas Costa <ncosta@codeaurora.org>