[qca-nss-ecm] Weak linkage between IP and MAC addresses so as to allow switching laptops that have same IP address quickly
When two laptops haave the same IP address and are swapped out on the lan quickly ECM remembers
the binding between the OLD MAC:IP and the newly connected MAC:IP.
When acceleration occurs this is done using the old MAC:IP and thus network traffic is disturbed.
This change creates a weak binding between MAC (ecm nodes) and IP addresses (ECM hosts) such that each
connection now keeps its own copy of the node that it is using.
This allows the same IP address to have multiple MAC's associated with it by virtue of the connection object that
records these informations.

CR 699045
CR 678434

Change-Id: I8c78675a56d762305740d76397077e9ba72d73b5
Signed-off-by: Gareth Williams <garethw@codeaurora.org>