[qca-nss-ecm] Remove move to head when ref walking hash chains
Remove code that move an object to the beginning of the hash chain.
Also remove ref-counted walk too.
The move to head is not efficient given the size of the hash and the
low incidence of collisions seen.
The ref-walking is slower than locking the chain given the overhead
of spinlocks and the ref counting operations at each step of the
iteration.

Change-Id: Id896f21bef34d8a4cb924edd7444c0b0b7974017
Signed-off-by: Gareth Williams <garethw@codeaurora.org>