[netlink] Updating netlink interface
This patch updates the netlink interface for kernel versions >= 3.13.
The API to registering a netlink family has changed from a series of
calls to a single call.
This call now takes the family to register, an array of ops, and an
array of groups (in the case of multicast) for registration.
In addition, op commands must either define the .cmd or .dumpit function
calls otherwise registration fails.

Change-Id: I0933521a379f67e89c919ce8cd8227717e44febe
Signed-off-by: Ryan Sherlock <rsherloc@codeaurora.org>
Signed-off-by: Xiaoping Fan <xfan@codeaurora.org>