[qca-nss-ecm] Added fix for acceleration not working for QinQ
Acceleration is not working for QinQ VLAN due to change in
vlan_dev_real_dev() Kernel API in 3.14 kernel. For 3.14 kernel
the vlan_dev_real_dev() always returns the physical dev in case
of QinQ VLAN. This is causing the ECM to push wrong rule to fw.
To fix this added a new kernel API(vlan_dev_next_dev) which returns
the next VLAN dev i.e.

eth0.1.10 -> eth0.1 -> eth0, The new API returns eth0.1 when
eth0.1.10 is passed as a parameter.

Change-Id: I1d8d232ea6ac382e5210e3c693a0b2ead2c24e81
Signed-off-by: Shyam Sunder <ssunde@codeaurora.org>