[qca-nss-ecm] Fix deadlock due to wrong nested lock calls
While the ECM frontend lock (e.g: ecm_nss_ipv4_lock) is held
we shouldn't hold the feci->lock, because in all over the ECM code,
the order of these lock calls are opposite. So, in ecm_db_connection_add
function, we need to remove the feci->ref() and call it outside of
ECM frontend lock.

Change-Id: I6dd9ecbbd37ef9a8308e27977f7e40c0f337b5d4
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>