[qca-nss-ecm] Added a NULL check for IPv6 gateway address
In some scenario (i.e DS-lite) the gateway address of a IPv6
flow is NULL and RTF_GATEWAY bit is not set if sourece and
destination address is in same network. In such cases check
if gateway address is NULL and do not overwrite the destination
IPv6 address used for dest MAC lookup.

Change-Id: Ic78e1397f6f44e39f1dd5729805ad71df74e4692
Signed-off-by: Shyam Sunder <ssunde@codeaurora.org>