[qca-nss-ecm] Update accel engine interface id
In some cases, the accel engine interface identifier
can be changed. In these cases, when the new connections
are being created, ECM does not establish a new interface
instance, because there is already one in the database with
the same MAC address and netdev index number. But since the
accel engine interface identifier is changed, we need to update
the interface instances ae_interface_identifier field.

Change-Id: Iadad8aa6e852151bc4615eb10eaf5363b29dde73
Signed-off-by: Murat Sezgin <msezgin@codeaurora.org>